+++ lib/runtime/server/ssr/index.ts
@@ -1,9 +1,13 @@
 export { WatchtowerEvents } from './render-events'
 export { RenderMarkup } from './render-app-to-string'
-export { default as renderPageContents, Assets, WatchtowerOptions } from './full-render'
+export {
+    default as renderPageContents, Assets, ServerSideRenderOptions, CreateReduxStore,
+} from './full-render'
 export {
     ServerRenderResult,
     ServerRenderResultType,
     SuccessServerRenderResult,
     PageNotFoundRenderResult,
 } from './server-render-results'
+export { default as recursiveTaskResolver } from './helpers/recursive-task-resolver'
+export * from './create-middleware'