+++ lib/runtime/server/ssr/render-events.ts
@@ -1,4 +1,5 @@
-import PromiseTracker from './lib/promise-tracker'
+import { PromiseTracker } from '../../universal'
+
 export interface WatchtowerEvents {
     beginWaitingForTasks?: (elapsed: string) => void
     // Hook to register additional work with the promise tracker after a render